Modulo-8_Testes/Feitos-Durante-a-Aula/Aula-2_Qua_29-06-2022/helper-api/src/services/SortService.ts
export enum TipoOrdenacao {
  Asc = 0,
  Desc = 1
}

export class SortService {
  getSort = (
    palavras: (string | number)[],
    removerDuplicados: boolean,
    ordenacao?: TipoOrdenacao
  ): (string | number)[] => {
    let result = [...palavras]

    if (ordenacao !== null && Object.values(TipoOrdenacao).includes(ordenacao!)) {
      result.sort()

      if (ordenacao === TipoOrdenacao.Desc)
        result.reverse()
    }

    if (removerDuplicados)
      result = [...new Set(result)]

    return result
  }
}